requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
requests==2.31.0
python-dotenv==1.0.0
facebook-sdk==3.1.0
instagram-private-api==1.6.0
tweepy==4.14.0
SQLAlchemy==2.0.23
python-telegram-bot==20.7
google-generativeai==0.3.2
schedule==1.2.1
pandas==2.1.4
numpy==1.26.2
beautifulsoup4==4.12.2
pytz==2023.3
Flask-WTF==0.15.1 